refactor api/lock.go
