+++ internal/api/cluster.go
@@ -2,7 +2,6 @@ package api
 
 import (
 	"net/http"
-	"sync"
 
 	"github.com/gorilla/mux"
 	"github.com/mattermost/mattermost-cloud/internal/model"
@@ -23,41 +22,6 @@ func initCluster(apiRouter *mux.Router, context *Context) {
 	clusterRouter.Handle("", addContext(handleRetryCreateCluster)).Methods("POST")
 	clusterRouter.Handle("/kubernetes/{version}", addContext(handleUpgradeCluster)).Methods("PUT")
 	clusterRouter.Handle("", addContext(handleDeleteCluster)).Methods("DELETE")
-}
-
-// lockCluster synchronizes access to the given cluster across potentially multiple provisioning
-// servers.
-func lockCluster(c *Context, clusterID string) (*model.Cluster, int, func()) {
-	cluster, err := c.Store.GetCluster(clusterID)
-	if err != nil {
-		c.Logger.WithError(err).Error("failed to query cluster")
-		return nil, http.StatusInternalServerError, nil
-	}
-	if cluster == nil {
-		return nil, http.StatusNotFound, nil
-	}
-
-	locked, err := c.Store.LockCluster(clusterID, c.RequestID)
-	if err != nil {
-		c.Logger.WithError(err).Error("failed to lock cluster")
-		return nil, http.StatusInternalServerError, nil
-	} else if !locked {
-		c.Logger.Error("failed to acquire lock for cluster")
-		return nil, http.StatusConflict, nil
-	}
-
-	unlockOnce := sync.Once{}
-
-	return cluster, 0, func() {
-		unlockOnce.Do(func() {
-			unlocked, err := c.Store.UnlockCluster(cluster.ID, c.RequestID, false)
-			if err != nil {
-				c.Logger.WithError(err).Errorf("failed to unlock cluster")
-			} else if unlocked != true {
-				c.Logger.Error("failed to release lock for cluster")
-			}
-		})
-	}
 }
 
 // handleGetClusters responds to GET /api/clusters, returning the specified page of clusters.
